Refactored feature - one class per file
