+++ source/Xunit.Gherkin.Quick/Feature.cs
@@ -1,12 +1,10 @@
-﻿using Gherkin;
-using Gherkin.Ast;
+﻿using Gherkin.Ast;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 using System.Text.RegularExpressions;
 using Xunit.Abstractions;
-using Xunit.Sdk;
 
 namespace Xunit.Gherkin.Quick
 {
@@ -86,87 +84,6 @@ namespace Xunit.Gherkin.Quick
                     throw;
                 }
             }
-        }
-    }
-
-    [AttributeUsage(AttributeTargets.Method, AllowMultiple = false)]
-    [XunitTestCaseDiscoverer("Xunit.Gherkin.Quick.ScenarioDiscoverer", "Xunit.Gherkin.Quick")]
-    internal sealed class ScenarioAttribute : FactAttribute
-    { }
-
-    [AttributeUsage(AttributeTargets.Class, AllowMultiple = false)]
-    public sealed class FeatureFileAttribute : Attribute
-    {
-        /// <summary>
-        /// Specify path to feature file (can be relative to output folder).
-        /// </summary>
-        /// <param name="path">
-        /// Path to feature file. Can be relative to output folder,
-        /// e.g. "./SampleFolder/Sample.feature"
-        /// </param>
-        public FeatureFileAttribute(string path)
-        {
-            Path = path ?? throw new ArgumentNullException(nameof(path));
-        }
-
-        public string Path { get; }
-    }
-
-    public sealed class ScenarioDiscoverer : IXunitTestCaseDiscoverer
-    {
-        private readonly IMessageSink _messageSink;
-
-        public ScenarioDiscoverer(IMessageSink messageSink)
-        {
-            _messageSink = messageSink;
-        }
-
-        public IEnumerable<IXunitTestCase> Discover(ITestFrameworkDiscoveryOptions discoveryOptions, ITestMethod testMethod, IAttributeInfo factAttribute)
-        {
-            var gherkinDocument = GetGherkinDocumentByType(testMethod.TestClass.Class.ToRuntimeType());
-
-            foreach (var scenario in gherkinDocument.Feature.Children)
-            {
-                yield return new ScenarioXUnitTestCase(_messageSink, testMethod, $"{gherkinDocument.Feature.Name} :: {scenario.Name}", new object[] { scenario.Name });
-            }
-        }
-
-        public static GherkinDocument GetGherkinDocumentByType(Type classType)
-        {
-            var fileName = classType.FullName;
-            fileName = fileName.Substring(fileName.LastIndexOf('.') + 1) + ".feature";
-
-            if (!System.IO.File.Exists(fileName))
-            {
-                var path = (classType.GetTypeInfo().GetCustomAttributes(typeof(FeatureFileAttribute))
-                    .FirstOrDefault() as FeatureFileAttribute)
-                    ?.Path;
-
-                if (path == null || !System.IO.File.Exists(path))
-                {
-                    throw new GherkinException($"Cannot find feature file `{fileName}` in the output root directory. If it's somewhere else, use {nameof(FeatureFileAttribute)} to specify file path.");
-                }
-
-                fileName = path;
-            }
-
-            var parser = new Parser();
-            var gherkinDocument = parser.Parse(fileName);
-            return gherkinDocument;
-        }
-    }
-
-    public sealed class ScenarioXUnitTestCase : XunitTestCase
-    {
-        [Obsolete]
-        public ScenarioXUnitTestCase()
-        {
-        }
-
-        public ScenarioXUnitTestCase(IMessageSink messageSink, ITestMethod testMethod, string scenarioName, object[] testMethodArguments = null)
-            : base(messageSink, TestMethodDisplay.Method, testMethod, testMethodArguments)
-        {
-            DisplayName = scenarioName;
         }
     }
 }
